+++ client/src/app/Pixi/MouseManager.ts
@@ -1,6 +1,6 @@
 import { PixiManager } from '../../api/PixiManager';
 import { PieceType } from '../../_types/global/GlobalTypes';
-import { PieceObject } from './PieceObject';
+import { PieceObject } from './Ships/PieceObject';
 import { BoardCoords, MoveAttack } from './@PixiTypes';
 import { compareBoardCoords, idxsIncludes } from './Utils/PixiUtils';
 import { Ship } from './Ships/Ship';
