+++ client/src/app/Pixi/Ships/Ship.ts
@@ -2,7 +2,7 @@ import * as PIXI from 'pixi.js';
 import { PixiManager } from '../../../api/PixiManager';
 import { VisiblePiece, PieceType } from '../../../_types/global/GlobalTypes';
 import { Wrapper } from '../PixiObject';
-import { PieceObject } from '../PieceObject';
+import { PieceObject } from './PieceObject';
 import { BoardCoords, CanvasCoords } from '../@PixiTypes';
 import { boardCoordsFromLoc } from '../Utils/PixiUtils';
 import { StatIcon, StatType, STATICON_W } from '../Utils/StatIcon';
